requirements.txt
pandas
scikit-learn
numpy
joblib
# Google Cloud & Vertex AI Libraries
google-cloud-aiplatform
google-cloud-storage
# Kubeflow Pipelines (KFP)
kfp
# Filesystem spec for GCS access by pandas
fsspec
gcsfs          